recognize slightly different trace for racket7 modules
Temporary workaround for racket7: While the racket7 expander has been
made to match the current expander for most cases, it uses an
immediately-only expansion (i.e., a `#f` stop list for `local-expand`)
to check whether a module's content expands to `#%module-begin`. This
commit adjust the macro stepper to handle that, but it doesn't do that
in the best way. Also, problems remain due to differences in top-level
expansion.