sync with expander, new macro hiding algorithm (#28)
This syncs the macro stepper with pending changes to the macro expander and fixes lots of little bugs (eg, handling of different kinds of lifts in different module passes). It replaces the macro hiding algorithm with one that tracks term identity through arm and disarm steps in addition to scope and property changes. The macro stepper UI now shows armed and tainted syntax.

Corresponds to racket/racket#2968